RolesController: Added role priority verification and its tests

+++ src/modules/users/controllers/roles/roles.controller.ts
@@ -1,4 +1,5 @@
 import {
+    BadRequestException,
     Controller,
     Delete,
     Get,
@@ -16,10 +17,17 @@ import { UseScopeGuard } from 'src/utils/decorators/UseScopeGuards.decorator';
 import { ScopeGuard } from 'src/utils/guards/scope/Scope.guard';
 import { IsAuthenticatedGuard } from 'src/utils/guards/session/IsAuthenticated.guard';
 import { RoleRelatedPermissions } from 'src/models/types/UserPermissions';
+import { AuthUser } from 'src/utils/decorators/AuthUser.decorator';
+import { User } from 'src/models/types/User';
+import RolesResolver from 'src/utils/helpers/RolesResolver';
+import { UsersService } from '../../services/users/users.service';
 
 @Controller('roles')
 export class RolesController {
-    constructor(private readonly rolesService: RolesService) {}
+    constructor(
+        private readonly rolesService: RolesService,
+        private readonly usersService: UsersService,
+    ) {}
 
     // Fetching
 
@@ -56,6 +64,7 @@ export class RolesController {
     // }
 
     // Assigning and removing users from roles
+    // TODO: Consider cases if somebody is assigning/removing his own role? (this same with permissions in the future)
 
     @UseGuards(ScopeGuard)
     @UseScopeGuard([{ roles: [UserRole.ADMIN] }, { permissions: [RoleRelatedPermissions.ASSIGN_ROLE] }])
@@ -63,7 +72,17 @@ export class RolesController {
     async addUserToRole(
         @Param('userId', ParseIntPipe) userId: number,
         @Param('roleName', new ParseEnumPipe(UserRole)) role: UserRole,
+        @AuthUser() user: User,
     ) {
+        // Verify if somebody is not trying to assign role higher or equal to his own role
+        if(RolesResolver.compareUsersRolesPriority(user, { roles: [role] }) <= 0) 
+            throw new BadRequestException('You cannot assign role with higher or equal priority to your own role');
+
+        const userToUpdate = await this.usersService.findById(userId);
+        if (userToUpdate.roles.length !== 0 && RolesResolver.compareUsersRolesPriority(user, userToUpdate) <= 0)
+            throw new BadRequestException('You cannot add role to user with a higher or equal role than yours');
+            
+
         return this.rolesService.addUserToRole(userId, role);
     }
 
@@ -73,7 +92,18 @@ export class RolesController {
     async removeUserFromRole(
         @Param('userId', ParseIntPipe) userId: number,
         @Param('roleName', new ParseEnumPipe(UserRole)) role: UserRole,
+        @AuthUser() user: User,
     ) {
+        
+        // Verify if somebody is not trying to assign role higher or equal to his own role
+        if(RolesResolver.compareUsersRolesPriority(user, { roles: [role] }) <= 0) 
+            throw new BadRequestException('You cannot remove role with higher or equal priority to your own role');
+
+        // Verify if somebody is not trying to remove role from user with more important role
+        const userToUpdate = await this.usersService.findById(userId);
+        if (userToUpdate.roles.length !== 0 && RolesResolver.compareUsersRolesPriority(user, userToUpdate) <= 0)
+            throw new BadRequestException('You cannot remove role from user with a higher or equal role than yours');
+
         return this.rolesService.removeUserFromRole(userId, role);
     }
 }
